+++ derive_builder/src/options/struct_mode.rs
@@ -79,6 +79,7 @@ impl From<OptionsBuilder<StructMode>> for (StructOptions, OptionsBuilder<FieldMo
             setter_vis: b.setter_vis,
             setter_into: b.setter_into,
             try_setter: b.try_setter,
+            field_vis: b.field_vis,
             default_expression: field_default_expression,
             no_std: b.no_std,
             mode: {